refactor: Move mocha tests to `test`
- NPM packages now do not contain test files
- Only use root test path `test/*.js` to avoid js in test/*/node_modules
